+++ contracts/strategies/ITupleTradingStrategy.sol
@@ -20,7 +20,7 @@ import "./ITradingStrategy.sol";
 interface ITupleTradingStrategy {
     function validateTuple(
         ITradingStrategy.Swap calldata swap,
-        uint256[] calldata balances,
+        uint128[] calldata balances,
         uint256 indexIn,
         uint256 indexOut
     ) external returns (bool, uint128);
